+++ app/src/main/java/com/example/freefood/activities/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.example.freefood;
+package com.example.freefood.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
