+++ app/src/main/java/com/example/freefood/adapters/PostInfoWindowAdapter.java
@@ -1,25 +1,23 @@
-package com.example.freefood;
+package com.example.freefood.adapters;
 
 import android.content.Context;
 import android.os.Build;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
-import android.widget.TextView;
 
 import androidx.annotation.RequiresApi;
 
 import com.bumptech.glide.Glide;
-import com.example.freefood.databinding.FragmentProfileBinding;
+import com.example.freefood.R;
 import com.example.freefood.databinding.InfoWindowPostBinding;
 import com.example.freefood.models.Post;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.model.Marker;
 
 import org.json.JSONException;
-import org.parceler.Parcels;
 
-import static com.example.freefood.Utils.containsJsontoString;
+import static com.example.freefood.utils.Utils.containsJsontoString;
 
 public class PostInfoWindowAdapter implements GoogleMap.InfoWindowAdapter {
