+++ app/src/main/res/layout/activity_expanded_image.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ExpandedImageActivity">
+    tools:context=".activities.ExpandedImageActivity">
 
 <!--    <include-->
 <!--        android:id="@+id/include"-->
