+++ app/src/main/res/layout/activity_signup.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SignupActivity">
+    tools:context=".activities.SignupActivity">
 
     <EditText
         android:id="@+id/etName"
