+++ app/src/main/res/layout/fragment_map.xml
@@ -147,7 +147,7 @@
         android:layout_height="0dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintTop_toBottomOf="@id/horizontalScrollView"
-        tools:context=".MainActivity"
+        tools:context=".activities.MainActivity"
         tools:layout="@layout/fragment_map" />
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
